requirements.txt
opik
comet_ml
litellm
python-dotenv
google-generativeai